[feat] remote 패키지 분리 및 repository 이름 변경
