+++ app/src/main/java/com/hamcoding/berryy/data/source/remote/KrxApiClient.kt
@@ -1,7 +1,7 @@
-package com.hamcoding.berryy.data.source
+package com.hamcoding.berryy.data.source.remote
 
 import com.hamcoding.berryy.BuildConfig
-import com.hamcoding.berryy.data.model.DetailResponse
+import com.hamcoding.berryy.data.model.SearchResponse
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
@@ -9,21 +9,20 @@ import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.http.GET
 import retrofit2.http.Query
 
-interface DetailApiClient {
-
-    @GET("getDiviInfo")
-    suspend fun getDetailList(
-        @Query("crno") companyCode: String,
-        @Query("serviceKey") apiKey: String = BuildConfig.DETAIL_API_KEY,
-        @Query("resultType") resultType: String = "JSON",
-    ) : DetailResponse
+interface KrxApiClient {
+    @GET("getItemInfo")
+    suspend fun getStockList(
+        @Query("likeItmsNm") name: String,
+        @Query("serviceKey") apiKey: String = BuildConfig.KRX_API_KEY,
+        @Query("resultType") resultType: String = "json"
+    ): SearchResponse
 
     companion object {
 
         private const val BASE_URL =
-            "http://apis.data.go.kr/1160100/service/GetStocDiviInfoService/"
+            "https://apis.data.go.kr/1160100/service/GetKrxListedInfoService/"
 
-        fun create(): DetailApiClient {
+        fun create(): KrxApiClient {
             val logger = HttpLoggingInterceptor().apply {
                 level = HttpLoggingInterceptor.Level.BODY
             }
@@ -37,7 +36,7 @@ interface DetailApiClient {
                 .client(client)
                 .addConverterFactory(GsonConverterFactory.create())
                 .build()
-                .create(DetailApiClient::class.java)
+                .create(KrxApiClient::class.java)
 
         }
     }
